requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Werkzeug==2.3.7

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21

# AI and Audio Processing
openai==0.28.1
openai-whisper==20230918
mistralai==0.0.8
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1

# Web forms and validation
WTForms==3.0.1
email-validator==2.0.0

# File handling
python-multipart==0.0.6

# Date and time handling
python-dateutil==2.8.2

# Environment variables
python-dotenv==1.0.0

# Security
bcrypt==4.0.1
cryptography==41.0.4

# Development tools
python-jose[cryptography]==3.3.0

# Production server
gunicorn==21.2.0

# Utilities
requests==2.31.0
celery==5.3.1
redis==4.6.0

# Testing
pytest==7.4.2
pytest-flask==1.2.0

# Linting and formatting
flake8==6.0.0
black==23.7.0
